catalog: extract defaultCatalogTableColumnsFunc to separate file
