+++ plugins/catalog/src/components/CatalogTable/CatalogTable.tsx
@@ -47,6 +47,7 @@ import React, { ReactNode, useMemo } from 'react';
 import { columnFactories } from './columns';
 import { CatalogTableColumnsFunc, CatalogTableRow } from './types';
 import { PaginatedCatalogTable } from './PaginatedCatalogTable';
+import { defaultCatalogTableColumnsFunc } from './defaultCatalogTableColumnsFunc';
 
 /**
  * Props for {@link CatalogTable}.
@@ -75,52 +76,6 @@ const refCompare = (a: Entity, b: Entity) => {
     });
 
   return toRef(a).localeCompare(toRef(b));
-};
-
-/** @public */
-export const defaultCatalogTableColumnsFunc: CatalogTableColumnsFunc = ({
-  filters,
-  entities,
-}) => {
-  const showTypeColumn = filters.type === undefined;
-
-  return [
-    columnFactories.createTitleColumn({ hidden: true }),
-    columnFactories.createNameColumn({ defaultKind: filters.kind?.value }),
-    ...createEntitySpecificColumns(),
-    columnFactories.createMetadataDescriptionColumn(),
-    columnFactories.createTagsColumn(),
-  ];
-
-  function createEntitySpecificColumns(): TableColumn<CatalogTableRow>[] {
-    const baseColumns = [
-      columnFactories.createSystemColumn(),
-      columnFactories.createOwnerColumn(),
-      columnFactories.createSpecTypeColumn({ hidden: !showTypeColumn }),
-      columnFactories.createSpecLifecycleColumn(),
-    ];
-    switch (filters.kind?.value) {
-      case 'user':
-        return [];
-      case 'domain':
-      case 'system':
-        return [columnFactories.createOwnerColumn()];
-      case 'group':
-      case 'template':
-        return [
-          columnFactories.createSpecTypeColumn({ hidden: !showTypeColumn }),
-        ];
-      case 'location':
-        return [
-          columnFactories.createSpecTypeColumn({ hidden: !showTypeColumn }),
-          columnFactories.createSpecTargetsColumn(),
-        ];
-      default:
-        return entities.every(entity => entity.metadata.namespace === 'default')
-          ? baseColumns
-          : [...baseColumns, columnFactories.createNamespaceColumn()];
-    }
-  }
 };
 
 /** @public */
