+++ plugins/catalog/src/components/CatalogTable/index.ts
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 
-export { CatalogTable, defaultCatalogTableColumnsFunc } from './CatalogTable';
+export { CatalogTable } from './CatalogTable';
+export { defaultCatalogTableColumnsFunc } from './defaultCatalogTableColumnsFunc';
 export type { CatalogTableProps } from './CatalogTable';
 export type { CatalogTableRow, CatalogTableColumnsFunc } from './types';
